Convert auth_token to use session
This patch should emulate the same behaviour of the existing auth token
middleware but using the new session object for requests and token
management.

There is one test removed in this patch which was to do with having the
auth token refreshed if it was stale. This is no longer handled by the
middleware but expected to be managed by the auth plugin.

This fixes the existing behaviour that if an admin_token was given and
was marked invalid then the middleware would fallback to using the
username and password provided. If an authentication method fails then
this is something that should be addressed not compensated for.

Blueprint: keystoneclient-auth-token
Change-Id: I1e7a4a63fd52ace92a1000d3f9330961c42e411f
Closes-Bug: #1307252