Plugin loading from config objects
Provide a pattern for auth plugins to load themselves from a config
object. The first user of this will be auth_token middleware however it
is not likely to be the only user.

By doing this in an exportable way we are defining a single config file
format for specifying how to load a plugin for all services.

Loading from CONF is just one of the ways plugins might be loaded.
The decision not to expose params as oslo.config options is on
purpose as there is cliff and other loading mechanisms besides
oslo.config we need to support.

Blueprint: plugin-params
Change-Id: I353b26a1ffc04a20666e76f5bd2f1e6d7c19a22d