Make tests run against original client and session
The managers have a bad habit of reaching back into the client and
making assumptions about the values that are saved there. These
assumptions are not always correct when we use the session object.

Test all the versioned managers against a client that was constructed
with the old method and with a session object.

Change-Id: I93a26db7ae7e4d887aa815108be71c72b4a1f2bb