Remove client creation via discovery
The assumption that the only use for Discovery was to create a client
has been a problem for a long time. The circular dependency it caused is
the reason for the _discover file and what is left in discover.py is
mostly just there to support the generic client discovery.

There are some issues with the way that the client discovery works.
It currently doesn't work properly with sessions and we need to
determine if we really want to do a version discovery check or just
proxy to the real object as other clients do.

For now remove the whole stinking mess.

We could now merge _discover.py into discover.py but we're going to
remove it soon.

Change-Id: Iea4721664af288ce4f798bee0be5d008281a33aa